Add :has pseudo-class support
See https://www.w3.org/TR/selectors-4/#has-pseudo.

In order to properly implement `:has`, we had to support passing multiple `count()` args (for each entry in a selector list, as supported by CSS 4 selectors), which is not supported OOB in XPath/XLinq. The custom CssContext provides resolution to the "fn:sum" (prefix used just to disambiguate with the XPath 1.0 sum) and just sums the args.

Fixes #12